libcap: depend on host-attr for the host
When a host-libcap is built, it is built without libattr support.
However, libcap does not define some symbols based on this flag. This
becomes a problem when a build-host tool is run (e.g., ls) which uses
these symbols gets the host-libcap library via LD_LIBRARY_PATH.

In order to define all of the symbols, add libattr support on the host.
The RAISE_SETFCAP setting must also be disabled since the required
permissions are not available during install.

Signed-off-by: Ben Boeckel <mathstuf@gmail.com>